Shifted error handling to Duke class.
