+++ src/main/java/Duke.java
@@ -4,7 +4,7 @@ public class Duke {
     public static void main(String[] args) {
         printText("\t Hello! I'm Duke\n\t What can I do for you?");
 
-        ToDoList list = new ToDoList();
+        TaskList list = new TaskList();
         Scanner scanner = new Scanner(System.in);
         String input = scanner.next();
 
@@ -14,36 +14,28 @@ public class Duke {
                     case "list":
                         printText(list.list());
                         break;
-                    case "mark": {
-                        int num = Integer.parseInt(scanner.nextLine().strip());
-                        printText(list.mark(num - 1));
-                    }
+                    case "mark":
+                        printText(list.mark(scanner.nextLine().strip()));
                         break;
-                    case "unmark": {
-                        int num = Integer.parseInt(scanner.nextLine().strip());
-                        printText(list.unMark(num - 1));
-                    }
-                    case "delete": {
-                        int num = Integer.parseInt(scanner.nextLine().strip());
-                        printText(list.delete(num - 1));
-                    }
+                    case "unmark":
+                        printText(list.unMark(scanner.nextLine().strip()));
+                    case "delete":
+                        printText(list.delete(scanner.nextLine().strip()));
                         break;
                     case "todo":
                         printText(list.add(TaskType.ToDos, scanner.nextLine().strip()));
                         break;
                     case "deadline":
-                        printText(list.add(TaskType.Deadlines, scanner.nextLine()));
+                        printText(list.add(TaskType.Deadlines, scanner.nextLine().strip()));
                         break;
                     case "event":
-                        printText(list.add(TaskType.Events, scanner.nextLine()));
+                        printText(list.add(TaskType.Events, scanner.nextLine().strip()));
                         break;
                     default:
                         throw new DukeException("\t OOPS!!! I'm sorry, but I don't know what that means :-(");
                 }
             } catch (DukeException dukeException){
                 printText(dukeException.getMessage());
-            } catch (NumberFormatException exception) {
-                printText("\t OOPS!!! Task number is invalid.");
             }
 
             input = scanner.next();
